Separate Python bindings into submodules
